+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Changed
+ - Update media player getter
+
 ## 1.1.1 - 2020-01-26
 ### Added
  - Raids auto leave feature